Got the first user story test case working. Many errors The database implementation has lots of errors in it such as not following constraints like unique session ids while inserting into the session table. There was also erros in the populate script so I commented it out so that it wasn't causing issues anymore. There was also many errors in the queries in the database implementation such as "delete * from <table>" with no where clause?(which probably truncates the table). Also (I didn't change it) but when you specifySessionRecurrence() it is implied in the name that you would be changing the recurrence state in a session already specified, not creating a new session. And if you create a new session you can insert the recurring column inside the single insert statement rather than having to insert a half empty row, then updating it straight away to insert another column.
I also changed the test as you were testing how many timeslots that had the specified sessionID, however the specifiySessionReccurrence method didn't update anything about timetableslots and only changed the reccurring field to be one-off. Now it tests to make sure that the row that was supposed to be changed was changed and it returns a correct test.

So it seems to work and the test works but the database functions will need to be altered to remove the errors in them